prometheus/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
psycopg2-binary==2.9.7
PyYAML==6.0.1
psutil==5.9.5
